trainer/requirements.txt
flask
flask-sqlalchemy
gunicorn
plumbum
psycopg[binary,pool]
